+++ mayan/apps/checkouts/links.py
@@ -8,8 +8,8 @@ from .icons import (
     icon_checkin_document, icon_checkout_document, icon_checkout_info
 )
 from .permissions import (
-    permission_document_checkin, permission_document_checkin_override,
-    permission_document_checkout, permission_document_checkout_detail_view
+    permission_document_check_in, permission_document_checkout,
+    permission_document_checkout_detail_view
 )
 
 
@@ -29,23 +29,32 @@ def is_not_checked_out(context):
         return True
 
 
-link_checkout_list = Link(
+link_document_checkout_list = Link(
     icon_class=icon_checkout_info, text=_('Checkouts'),
-    view='checkouts:checkout_list'
+    view='checkouts:document_checkout_list'
 )
-link_checkout_document = Link(
-    args='object.pk', condition=is_not_checked_out,
-    icon_class=icon_checkout_document,
+link_document_checkout = Link(
+    condition=is_not_checked_out, icon_class=icon_checkout_document,
+    kwargs={'document_id': 'object.pk'},
     permission=permission_document_checkout, text=_('Check out document'),
-    view='checkouts:checkout_document',
+    view='checkouts:document_checkout',
 )
-link_checkin_document = Link(
-    args='object.pk', condition=is_checked_out,
-    icon_class=icon_checkin_document, permission=permission_document_checkin,
-    text=_('Check in document'), view='checkouts:checkin_document',
+link_document_multiple_checkout = Link(
+    icon_class=icon_checkout_document,
+    permission=permission_document_checkout, text=_('Check out'),
+    view='checkouts:document_multiple_checkout',
 )
-link_checkout_info = Link(
-    args='resolved_object.pk', icon_class=icon_checkout_info,
+link_document_check_in = Link(
+    condition=is_checked_out, icon_class=icon_checkin_document,
+    kwargs={'document_id': 'object.pk'}, permission=permission_document_check_in,
+    text=_('Check in document'), view='checkouts:document_check_in',
+)
+link_document_multiple_check_in = Link(
+    icon_class=icon_checkin_document, permission=permission_document_check_in,
+    text=_('Check in'), view='checkouts:document_multiple_check_in',
+)
+link_document_checkout_info = Link(
+    icon_class=icon_checkout_info, kwargs={'document_id': 'resolved_object.pk'},
     permission=permission_document_checkout_detail_view,
-    text=_('Check in/out'), view='checkouts:checkout_info',
+    text=_('Check in/out'), view='checkouts:document_checkout_info',
 )