+++ mayan/apps/checkouts/models.py
@@ -4,7 +4,7 @@ import logging
 
 from django.conf import settings
 from django.core.exceptions import ValidationError
-from django.db import models
+from django.db import models, transaction
 from django.urls import reverse
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.timezone import now
@@ -68,13 +68,14 @@ class DocumentCheckout(models.Model):
             )
 
     def delete(self, *args, **kwargs):
-        # TODO: enclose in transaction
-        NewVersionBlock.objects.unblock(self.document)
-        super(DocumentCheckout, self).delete(*args, **kwargs)
+        with transaction.atomic():
+            NewVersionBlock.objects.unblock(document=self.document)
+            super(DocumentCheckout, self).delete(*args, **kwargs)
 
     def get_absolute_url(self):
         return reverse(
-            viewname='checkout:checkout_info', kwargs={'pk': self.document.pk}
+            viewname='checkout:checkout_info',
+            kwargs={'document_id': self.document.pk}
         )
 
     def natural_key(self):
@@ -82,25 +83,27 @@ class DocumentCheckout(models.Model):
     natural_key.dependencies = ['documents.Document']
 
     def save(self, *args, **kwargs):
-        # TODO: enclose in transaction
         new_checkout = not self.pk
         if not new_checkout or self.document.is_checked_out():
-            raise DocumentAlreadyCheckedOut
-
-        result = super(DocumentCheckout, self).save(*args, **kwargs)
-        if new_checkout:
-            event_document_check_out.commit(
-                actor=self.user, target=self.document
-            )
-            if self.block_new_version:
-                NewVersionBlock.objects.block(self.document)
-
-            logger.info(
-                'Document "%s" checked out by user "%s"',
-                self.document, self.user
+            raise DocumentAlreadyCheckedOut(
+                _('Document already checked out.')
             )
 
-        return result
+        with transaction.atomic():
+            result = super(DocumentCheckout, self).save(*args, **kwargs)
+            if new_checkout:
+                event_document_check_out.commit(
+                    actor=self.user, target=self.document
+                )
+                if self.block_new_version:
+                    NewVersionBlock.objects.block(self.document)
+
+                logger.info(
+                    'Document "%s" checked out by user "%s"',
+                    self.document, self.user
+                )
+
+            return result
 
 
 class NewVersionBlock(models.Model):