+++ HISTORY.rst
@@ -225,6 +225,8 @@
 - Remove the permissions to grant or revoke a permission to a role.
   The instead the role edit permission is used.
 - Add a test mixin to generate random model primary keys.
+- Add support for checkout and check in multiple documents at
+  the same time.
 
 3.1.9 (2018-11-01)
 ==================